requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.7.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.5
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.7.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
httplib2==0.31.0
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.14.0
PyMuPDF==1.26.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9.1
safetensors==0.5.3
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
